refactor: Renamed app folder to api
